Update Interchange_Simulation.py
edited prodSimulation method so as to have separate variables for how frequently to save to the log file and how frequently to save to the trajectory. From testing OpenMM runs considerably slower when frequently saving to the log file (maybe it is how detailed we request the log file output to be), so this is in an effort to speed it all up.

Next to test is splitting the simulation itself into multiple calls to see if this also speeds it up for us. There's an issue on our servers where as a job proceeds the utilization of gpu's steadily decreases...I'm trying to circumvent this so as to increase our run speeds